refactor(api):修改UID的生成方式, 并将于将来替换掉其他的生成` (#5)
* refactor(api):修改Uid的生成方式, 变为8位int

* fix(api):修复c85613808a0a2e9ed2ab4120b8a1b51486888b71创建的死循环, 修复UID位数问题

* chore(web):允许本地调试

* style(web):format

+++ api/src/main/kotlin/cn/hirrao/javaee/entity/User.kt
@@ -5,7 +5,7 @@ import com.baomidou.mybatisplus.annotation.TableId
 
 //JsonIgnore注解是jackson包中的注解，用于返回json数据时，忽略该字段
 data class User(
-    @TableId(value = "uid") var uid: Long,
+    @TableId(value = "uid") var uid: Int,
 
     @TableField("userName") var userName: String? = null,
