+++ api/src/main/kotlin/cn/hirrao/javaee/mapper/DrugAlertMapper.kt
@@ -12,7 +12,7 @@ import java.time.LocalTime
 @Mapper
 interface DrugAlertMapper : BaseMapper<DrugAlert?> {
     @Insert("INSERT INTO drugAlert VALUES (#{alertId},#{uid},#{drugId}, #{alertTime},#{eatTime})")
-    fun insertDrugAlert(alertId: Long, uid: Long, drugId: Long, alertTime: LocalTime?, eatTime: LocalDate?)
+    fun insertDrugAlert(alertId: Long, uid: Int, drugId: Long, alertTime: LocalTime?, eatTime: LocalDate?)
 
     @Update("UPDATE drug SET isActive=#{isActive} WHERE alertId=#{alertId}")
     fun updateDrugAlertIsActiveById(alertId: Long, isActive: Int)
@@ -21,5 +21,5 @@ interface DrugAlertMapper : BaseMapper<DrugAlert?> {
     fun updateDrugAlertEatTimeById(alertId: Long, eatTime: LocalDate?)
 
     @Delete("DELETE FROM drugAlert WHERE alertId=#{alertId} AND uid=#{uid} AND drugId=#{drugId}")
-    fun deleteDrugAlertById(alertId: Long, uid: Long, drugId: Long)
+    fun deleteDrugAlertById(alertId: Long, uid: Int, drugId: Long)
 }
