+++ api/src/main/kotlin/cn/hirrao/javaee/service/UserIntroService.kt
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service
 
 @Service
 interface UserIntroService {
-    fun findUserIntro(uid: Long): UserIntro?
+    fun findUserIntro(uid: Int): UserIntro?
     fun updateUserIntro(uid: Long, intro: String?)
-    fun setUserIntro(uid: Long, intro: String?)
+    fun setUserIntro(uid: Int, intro: String?)
 }
